refactor (accounts): change controller into handler and move interface into models
