--- features/accounts/controller/account_controller.go
@@ -1,9 +1,0 @@
-package controller
-
-import "github.com/labstack/echo/v4"
-
-type AccountController interface {
-	Create(c echo.Context) error
-	GetDetail(c echo.Context) error
-	GetList(c echo.Context) error
-}
